mimirpb: add FromLabelAdaptersToKeyString
Avoid converting to labels and then to string.
